+++ examples/video_call/main/app_main.c
@@ -63,8 +63,8 @@ void app_main(void)
   init_config_t init_cfg = {
       .meetingID = CONFIG_VIDEOSDK_MEETING_ID,
       .token = token,
-      .displayName = "ESP32S3-Video", // user configuraable for display in the meeting
-      .participantId = deviceid,      
+      .displayName = "ESP32S3-Video", // any name you like; shown in the meeting
+      .participantId = deviceid,      // this device's id
       .audioCodec = AUDIO_CODEC_PCMA,
       .videoCodec = VIDEO_CODEC_JPEG,
   };